AddIP Methods at AddIP.java
