+++ src/main/java/de/ghc/managementbot/commands/UpdateIP.java
@@ -5,7 +5,11 @@ import de.ghc.managementbot.content.Data;
 import de.ghc.managementbot.entity.Command;
 import de.ghc.managementbot.entity.IPEntry;
 import de.ghc.managementbot.threads.DeleteMessageThread;
+import net.dv8tion.jda.core.entities.User;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
+
+import java.util.HashMap;
+import java.util.Map;
 
 public class UpdateIP implements Command {
     private IPEntry entry;
@@ -21,6 +25,18 @@ public class UpdateIP implements Command {
         if (msg.equalsIgnoreCase("ja") || msg.equalsIgnoreCase("j") || msg.equalsIgnoreCase("y")) {
             Content.getGhc().getTextChannelById(Data.Channel.zahlenschlacht).sendMessage(entry.toString() + "Daten von: " + Content.getGHCMember(entry.getAddedBy()).getEffectiveName()).queue();
         }
-        Content.deleteUserUpdateIP(event.getAuthor(), this);
+        deleteUserUpdateIP(event.getAuthor(), this);
+    }
+    private static Map<User, UpdateIP> userUpdateIP = new HashMap<>();
+
+    public static Map<User, UpdateIP> getUpdateIP() {
+        return userUpdateIP;
+    }
+
+    public static void addUserUpdateIP(User user, UpdateIP updateIP) {
+        userUpdateIP.put(user, updateIP);
+    }
+    public static void deleteUserUpdateIP(User user, UpdateIP update) {
+        userUpdateIP.remove(user, update);
     }
 }
